Negative binomial branching process model

Negative binomial branching process model

run_sim(
  n,
  n_length,
  mean_si,
  sd_si,
  R0,
  k = 0.16,
  tf = 37,
  max_potential_cases,
  delay_sample
)

Arguments

n
Numeric, number of initial cases

n_length
Numeric, the number of days in the initial seeding event

mean_si
Numeric, the mean of the serial interval

sd_si
Numeric, the sd of the serial interval

R0
Numeric, the estimated reproduction number

k
Numeric, the dispersion of the negative binomial branching process

tf
Numeric, the end time of the branching process

max_potential_cases
Numeric, the maximum number of cases

delay_sample
A function to sample from reporting delays

Value

A dataframe containing the simulation time and outbreak size

Examples

run_sim(n = 1, n_length = 7, mean_si = 5, sd_si = 2, R0 = 2,
        k=0.16, tf=37, max_potential_cases= 100,
        delay_sample = function(x) {rnorm(x, 6, 1)})

#>    time size
#> 1     0    0
#> 2     1    0
#> 3     2    0
#> 4     3    0
#> 5     4    0
#> 6     5    0
#> 7     6    0
#> 8     7    0
#> 9     8    0
#> 10    9    0
#> 11   10    0
#> 12   11    0
#> 13   12    1
#> 14   13    1
#> 15   14    1
#> 16   15    1
#> 17   16    1
#> 18   17    1
#> 19   18    1
#> 20   19    1
#> 21   20    1
#> 22   21    1
#> 23   22    1
#> 24   23    1
#> 25   24    1
#> 26   25    1
#> 27   26    1
#> 28   27    1
#> 29   28    1
#> 30   29    1
#> 31   30    1
#> 32   31    1
#> 33   32    1
#> 34   33    1
#> 35   34    1
#> 36   35    1
#> 37   36    1
#> 38   37    1

## Code
run_sim

#> function (n, n_length, mean_si, sd_si, R0, k = 0.16, tf = 37, 
#>     max_potential_cases, delay_sample) 
#> {
#>     t0 <- sample(seq(1, n_length), n, replace = TRUE)
#>     sim <- bpmodels::chain_sim(n, "nbinom", serial = function(x) stats::rnorm(x, 
#>         mean_si, sd_si), mu = R0, size = k, t0 = t0, tf = tf, 
#>         infinite = max_potential_cases)
#>     sim$time <- sim$time + delay_sample(length(sim$time))
#>     sim <- sim[sim$time < tf, ]
#>     size <- data.frame(time = 0:tf, size = unlist(lapply(0:tf, 
#>         function(x) {
#>             tmp <- sim[sim$time < x, ]
#>             nrow(tmp)
#>         })))
#>     return(size)
#> }
#> <bytecode: 0x55d8a64c60b8>
#> <environment: namespace:WuhanSeedingVsTransmission>